requirements.txt
fastapi==0.95.2
Jinja2==3.1.3
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
pydantic==1.10.13
qdrant-client==1.4.0
typing-extensions==4.5.0
Requests==2.31.0
seaborn==0.13.2
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.31.0
huggingface-hub==0.19.4
tqdm==4.66.1
uvicorn==0.27.0.post1
fastparquet
pyarrow
python-multipart
#python 3.10.12